mtd: mtdblock: fix nested sleeps in erase_write
Source: linux-mvista-2.4
MR: 95115, 00000
Type: Integration
Disposition: Merged from linux-mvista-2.4
ChangeID: 14b187e890c0995e58402e217a10e77ec819db86
Description:

When mtdblock is used with a device that can sleep,
this wait-for-callback in erase_write can lose a wakeup,
and/or provoke a warning
	"do not call blocking ops when !TASK_RUNNING"
See https://lwn.net/Articles/628628/ for background.

This was discovered with the staging mt29f_spinand driver,
which blocks on SPI bus mutex.
But the problem is here, where the old code assumed
underlying device would not sleep.
This is a generic layer that should work across all MTD,
so lose that assumption.

Signed-off-by: Peter Swain <pswain@cavium.com>
Signed-off-by: Corey Minyard <cminyard@mvista.com>
Signed-off-by: Jeremy Puhlman <jpuhlman@mvista.com>